refactor(#193): remove barrel pattern
