--- apps/bff/src/common/constants/index.ts
@@ -1,2 +1,0 @@
-export * from './error-messages.js';
-export * from './http-status-maps.js';
