--- apps/bff/src/common/filters/index.ts
@@ -1,1 +1,0 @@
-export * from './http-exception.filter.js';
