--- apps/bff/src/common/index.ts
@@ -1,4 +1,0 @@
-export * from './constants/index.js';
-export * from './filters/index.js';
-export * from './interceptors/index.js';
-export * from './types/index.js';
