--- apps/bff/src/common/interceptors/index.ts
@@ -1,1 +1,0 @@
-export * from './logging.interceptor.js';
