--- apps/bff/src/common/types/index.ts
@@ -1,1 +1,0 @@
-export * from './api-response.types.js';
