+++ apps/bff/src/main.ts
@@ -2,7 +2,8 @@ import { NestFactory } from '@nestjs/core';
 import { Logger } from '@nestjs/common';
 
 import { AppModule } from './app.module.js';
-import { HttpExceptionFilter, LoggingInterceptor } from './common/index.js';
+import { HttpExceptionFilter } from './common/filters/http-exception.filter.js';
+import { LoggingInterceptor } from './common/interceptors/logging.interceptor.js';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
